+++ apps/space/components/issues/navbar/issue-filter.tsx
@@ -19,27 +19,25 @@ export const NavbarIssueFilter = observer(() => {
   const pathName = router.asPath;
 
   const handleOnSelect = (key: "states" | "labels" | "priorities", value: string) => {
-    if (key === "states") {
-      store.issue.userSelectedStates = store.issue.userSelectedStates.includes(value)
-        ? store.issue.userSelectedStates.filter((s) => s !== value)
-        : [...store.issue.userSelectedStates, value];
-    } else if (key === "labels") {
-      store.issue.userSelectedLabels = store.issue.userSelectedLabels.includes(value)
-        ? store.issue.userSelectedLabels.filter((l) => l !== value)
-        : [...store.issue.userSelectedLabels, value];
-    } else if (key === "priorities") {
-      store.issue.userSelectedPriorities = store.issue.userSelectedPriorities.includes(value)
-        ? store.issue.userSelectedPriorities.filter((p) => p !== value)
-        : [...store.issue.userSelectedPriorities, value];
-    }
-
-    const paramsCommaSeparated = `${`board=${store.issue.currentIssueBoardView || "list"}`}${
-      store.issue.userSelectedPriorities.length > 0 ? `&priorities=${store.issue.userSelectedPriorities.join(",")}` : ""
-    }${store.issue.userSelectedStates.length > 0 ? `&states=${store.issue.userSelectedStates.join(",")}` : ""}${
-      store.issue.userSelectedLabels.length > 0 ? `&labels=${store.issue.userSelectedLabels.join(",")}` : ""
-    }`;
-
-    router.replace(`${pathName}?${paramsCommaSeparated}`);
+    // if (key === "states") {
+    //   store.issue.userSelectedStates = store.issue.userSelectedStates.includes(value)
+    //     ? store.issue.userSelectedStates.filter((s) => s !== value)
+    //     : [...store.issue.userSelectedStates, value];
+    // } else if (key === "labels") {
+    //   store.issue.userSelectedLabels = store.issue.userSelectedLabels.includes(value)
+    //     ? store.issue.userSelectedLabels.filter((l) => l !== value)
+    //     : [...store.issue.userSelectedLabels, value];
+    // } else if (key === "priorities") {
+    //   store.issue.userSelectedPriorities = store.issue.userSelectedPriorities.includes(value)
+    //     ? store.issue.userSelectedPriorities.filter((p) => p !== value)
+    //     : [...store.issue.userSelectedPriorities, value];
+    // }
+    // const paramsCommaSeparated = `${`board=${store.issue.currentIssueBoardView || "list"}`}${
+    //   store.issue.userSelectedPriorities.length > 0 ? `&priorities=${store.issue.userSelectedPriorities.join(",")}` : ""
+    // }${store.issue.userSelectedStates.length > 0 ? `&states=${store.issue.userSelectedStates.join(",")}` : ""}${
+    //   store.issue.userSelectedLabels.length > 0 ? `&labels=${store.issue.userSelectedLabels.join(",")}` : ""
+    // }`;
+    // router.replace(`${pathName}?${paramsCommaSeparated}`);
   };
 
   return (
@@ -69,7 +67,7 @@ export const NavbarIssueFilter = observer(() => {
               </span>
             ),
             onClick: () => handleOnSelect("priorities", priority),
-            isSelected: store.issue.userSelectedPriorities.includes(priority),
+            isSelected: store.issue.filteredPriorities.includes(priority),
           })),
         },
         {
@@ -85,7 +83,7 @@ export const NavbarIssueFilter = observer(() => {
                 </span>
               ),
               onClick: () => handleOnSelect("states", state.id),
-              isSelected: store.issue.userSelectedStates.includes(state.id),
+              isSelected: store.issue.filteredStates.includes(state.id),
             };
           }),
         },
@@ -104,7 +102,7 @@ export const NavbarIssueFilter = observer(() => {
               </span>
             ),
             onClick: () => handleOnSelect("labels", label.id),
-            isSelected: store.issue.userSelectedLabels.includes(label.id),
+            isSelected: store.issue.filteredLabels.includes(label.id),
           })),
         },
       ]}
