+++ apps/space/components/issues/peek-overview/issue-details.tsx
@@ -1,7 +1,7 @@
 // components
 import { IssueReactions } from "components/issues/peek-overview";
 // types
-import { IIssue } from "store/types";
+import { IIssue } from "types";
 
 type Props = {
   issue: IIssue;
