+++ apps/space/components/views/project-details.tsx
@@ -1,3 +1,4 @@
+import { useEffect } from "react";
 import { useRouter } from "next/router";
 import { observer } from "mobx-react-lite";
 // components
@@ -10,15 +11,16 @@ import { IssuePeekOverview } from "components/issues/peek-overview";
 // mobx store
 import { RootStore } from "store/root";
 import { useMobxStore } from "lib/mobx/store-provider";
-import { useEffect } from "react";
 
-export const ProjectDetailsView = () => {
+export const ProjectDetailsView = observer(() => {
   const router = useRouter();
   const { workspace_slug, project_slug, states, labels, priorities } = router.query;
 
-  const { issue: issueStore }: RootStore = useMobxStore();
+  const { issue: issueStore, project: projectStore }: RootStore = useMobxStore();
 
-  const activeIssueId = issueStore.activePeekOverviewIssueId;
+  console.log("projectStore?.activeBoard", projectStore?.activeBoard);
+
+  const activeIssueId = issueStore.peekId;
 
   useEffect(() => {
     if (workspace_slug && project_slug) {
@@ -33,14 +35,14 @@ export const ProjectDetailsView = () => {
 
   return (
     <div className="relative w-full h-full overflow-hidden">
-      {workspace_slug && (
+      {/* {workspace_slug && (
         <IssuePeekOverview
           isOpen={Boolean(activeIssueId)}
           onClose={() => issueStore.setActivePeekOverviewIssueId(null)}
           issue={issueStore?.issues?.find((_issue) => _issue.id === activeIssueId) || null}
           workspaceSlug={workspace_slug.toString()}
         />
-      )}
+      )} */}
 
       {issueStore?.loader && !issueStore.issues ? (
         <div className="text-sm text-center py-10 text-custom-text-100">Loading...</div>
@@ -51,23 +53,23 @@ export const ProjectDetailsView = () => {
               Something went wrong.
             </div>
           ) : (
-            issueStore?.currentIssueBoardView && (
+            projectStore?.activeBoard && (
               <>
-                {issueStore?.currentIssueBoardView === "list" && (
+                {projectStore?.activeBoard === "list" && (
                   <div className="relative w-full h-full overflow-y-auto">
                     <div className="mx-auto px-4">
                       <IssueListView />
                     </div>
                   </div>
                 )}
-                {issueStore?.currentIssueBoardView === "kanban" && (
+                {projectStore?.activeBoard === "kanban" && (
                   <div className="relative w-full h-full mx-auto px-9 py-5">
                     <IssueKanbanView />
                   </div>
                 )}
-                {issueStore?.currentIssueBoardView === "calendar" && <IssueCalendarView />}
-                {issueStore?.currentIssueBoardView === "spreadsheet" && <IssueSpreadsheetView />}
-                {issueStore?.currentIssueBoardView === "gantt" && <IssueGanttView />}
+                {projectStore?.activeBoard === "calendar" && <IssueCalendarView />}
+                {projectStore?.activeBoard === "spreadsheet" && <IssueSpreadsheetView />}
+                {projectStore?.activeBoard === "gantt" && <IssueGanttView />}
               </>
             )
           )}
@@ -75,4 +77,4 @@ export const ProjectDetailsView = () => {
       )}
     </div>
   );
-};
+});
