+++ apps/space/types/index.ts
